hotel_agent/requirement.txt
fastapi
uvicorn
python-multipart
chromadb
pypdf
python-docx
markdown
beautifulsoup4
httpx
openai
python-dotenv
pydantic

